apps/backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
boto3==1.34.162
psycopg2-binary==2.9.9
redis==5.0.7
python-dotenv==1.0.1
loguru==0.7.2
pdfminer.six==20231228
lxml==5.3.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.4.0
reportlab==4.2.0
openpyxl==3.1.2
